hello-slack/requirements.txt
slack_bolt
fastapi
uvicorn
python-dotenv 